PointSampler

PointSampler.bad_indices

PointSampler.discard_bad_seeds!

PointSampler.distance

PointSampler.good_indices

PointSampler.monte_carlo_th

PointSampler.monte_carlo_th

PointSampler.projected_distance

PointSampler.projected_distance

PointSampler.projected_distance_thresholded

PointSampler.bad_indices — Function

Indices of vectors in set that violate the box constraints lb or ub.

PointSampler.discard_bad_seeds! — Function

Delete vectors from array seeds that violate the box constraints lb and ub.

PointSampler.distance — Method

Return Euclidean distance from p1 to every point in P.

PointSampler.good_indices — Function

Indices of vectors in set that conform to the box constraints lb or ub.

PointSampler.monte_carlo_th — Function

Scale the design returned by the unconstrained version of this function to the box defined by lb and ub.

PointSampler.monte_carlo_th — Method

montecarloth( npoints = 10, ndims = 2; seeds = [], spawnfactor = 50, pdistthreshold_tolerance = 0.5 )

Return an array of length n_points containing real vectors representing points in space with n_dims dimensions. The points are iteratively chosen from random point sets to maximize a space-filling criterion as described in

"Surrogate Modelling of Computer Experiments with Sequential Experimental Design.", Crombecq, 2011

The returned point set is constructed starting with the points in seeds. If seeds is empty (default), then the singleton set containing the zero vector is used.

PointSampler.projected_distance — Method

Return the minium projected distance between two point vectors p1 and p2, i.e. the minimum absolute value of the differences of the coordinates of p1 and p2.

PointSampler.projected_distance — Method

Return array of projected distances for point p1 against every point in P.

PointSampler.projected_distance_thresholded — Function

Return array of projected distances for point p1 against every point in P, but every value below threshold is set to 0.0.